requirements.txt
pandas
mysql-connector-python
matplotlib
seaborn
